requirements.txt
dnspython
OSMPythonTools

